Setup class introduced
With the Setup class we can pass all the parameters needed for the Twitter query (dev.account, keywords, languages...) by simply filling a form. By pressing the launch button, the DataDownloader jar will be launched. TwitterManager and DataDownloader have been modified in  order to work with the new Setup class.